+++ app/src/main/java/fr/didictateur/inanutshell/RecetteDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 @Dao
 public interface RecetteDao {
 	@Query("SELECT * FROM Recette Where parentId IS :parentId")
-	List<Recette> getRecettesByParent(Integer parentId);
+	List<Recette> getRecettesByParent(Long parentId);
 
 	@Insert
 	long insert(Recette recette);
@@ -18,7 +18,7 @@ public interface RecetteDao {
 	void delete(Recette recette);
 
 	@Query("SELECT * FROM Recette WHERE id = :id")
-	Recette getRecetteById(int id);
+	Recette getRecetteById(long id);
 
 	@Query("SELECT * FROM Recette")
 	List<Recette> getAllRecettes();
